+++ yellowbox/__init__.py
@@ -1,6 +1,6 @@
 from yellowbox._version import __version__
 from yellowbox.clients import docker_client
-from yellowbox.image_build import image_build
+from yellowbox.image_build import build_image
 from yellowbox.networks import connect, temp_network
 from yellowbox.retry import RetrySpec
 from yellowbox.service import YellowService
@@ -9,7 +9,7 @@ from yellowbox.subclasses import ContainerService, RunMixin, SingleContainerServ
 __all__ = [
     '__version__',
     'docker_client',
-    'image_build',
+    'build_image',
     'connect', 'temp_network',
     'RetrySpec',
     'YellowService',